refactor(step2): 테스트를 클래스별로 분리
